+++ src/Domain/Entities/Resposta.cs
@@ -3,6 +3,6 @@ public class Resposta : BaseAuditableEntity
 {
     public required string RespostaQuestionario { get; set;}
     public required Questionario Questionario { get; set; }
-    public required int ValorPesoResposta { get; set; }
+    public required decimal ValorPesoResposta { get; set; }
 
 }
